add 'payload' property to images
This is to solve a problem we have with the Fedora composes.
There are 'Spins' and 'Labs' variants which produce several
live images and ARM appliances with differing content. With the
current metadata, there is no good way to tell the KDE live
image from the Xfce or LXDE live images; the only property
which indicates this is the 'path', but that has lots of other
information in it and is not trivially parseable (as image
names don't really strictly follow the naming policy you
cannot strictly tell what element of the image file name is
the 'payload').

Thus we add a 'payload' property which Pungi will populate as
appropriate. For many images the 'payload' will be the variant,
but for Spins and Labs it will indicate the actual content of
each image.

The image metadata version should probably change with this,
but I couldn't see the right way to do that (surely I don't
just universally bump it to 1, 1 in common.py?) Pungi will
also need a few corresponding changes, I will send a Pagure
PR for those.