Add __repr__ methods to more classes
Specifically `Variant`, `Compose`, `VariantPath`, `BaseProduct`
(inherited in `Release`) and `VariantBase`.

This is helpful during development as it makes debug outputs actually
usable. There is only some very basic information, but it is better than
showing a memory address. Due to how these classes work the output can
not be copy-pasted into REPL to create the object.