add __version__ attribute and drop hard-coded version
This attribute is described at https://www.python.org/dev/peps/pep-0396/

This allows productmd consumers to report the exact version of productmd
that is in use, or conditionalize consumer behavior based on this
version number, etc.

We can use this attribute in setup.py, rather than hard-coding it there.