composeinfo: py3 support for sort in get_variants()
In Python 2, list.sort() takes a cmp parameter. This parameter is a
function which does the comparison between two elements.

In Python 3, the cmp parameter was removed entirely. This causes the
following error in get_variants() on py3:

  TypeError: must use keyword argument for key function

Simplify the sorting operation with the key kwarg to be compatible with
both py2 and py3.