Support 'unique image identifier' concept, enforce on 1.1+
I've been doing this sort of ad hoc in Fedora QA tools for a
while; per discussion in https://pagure.io/pungi/issue/525 , it
seemed a good idea to add the concept to productmd. This adds a
constant that declares the combination of subvariant, type,
format, arch and disc number for an image should always be
unique within a compose, allowing us to use it to identify 'the
same' image across multiple composes. An `identify_image`
function is added to `images` that provides a tuple of the
values.

Perhaps more controversially, this is also enforced in the
Images `add` method; adding an image with different checksums
from an existing image but the same unique attributes will raise
a ValueError.