Sync spec file with Fedora (enable Py3 for EPEL) (#79)
I made these changes 'downstream' in Fedora SCM without knowing
the spec was maintained here. This should sync the changes in a
way that works for everyone, I hope. By using the `%epel` macro
instead of `%rhel`, the Python 3 builds should not be enabled
when building for pure RHEL. `%epel` is available in all EPEL
branches since April - see RHBZ #1320497.

The Python macro compat cruft is also preserved (I originally
removed it downstream) because apparently it was only added to
EPEL, not to RHEL proper :( See RHBZ #1307190.

Added a comment to explain that the spec also exists in upstream
git, and is used for pure-RHEL builds.

Signed-off-by: Adam Williamson <awilliam@redhat.com>