Add a compose ID parser, tweak short and version rules
This adds a `parse_compose_id` function to `composeinfo`. It
can parse most compose IDs back into their component parts. As
part of this, we tweak the rules for short names and versions
a little. The `RELEASE_SHORT_RE` regex now allows upper-case
characters, as Fedora has been using these in short names for
some time (only `create_release_id` actually validated short
names, up until now, which is why we hadn't noticed this
inconsistency). The `RELEASE_VERSION_RE` regex now does *not*
allow absolutely any string that doesn't start with a digit to
be used as a version; the only string version allowed is
'Rawhide' (this is the only one that exists in the wild so far
as I know). This is to make it possible to write a sane parser
at all.

We also fix `get_date_type_respin` to work with the new `ci`
type.

We could possibly go further than this and disallow the use of
dashes in short names; if we did that, parsing compose IDs
would become much less difficult and slightly more reliable.
But it would require us to either rename or 'grandfather in'
Fedora's existing short names with dashes in them, which are
Fedora-Atomic, Fedora-Docker and Fedora-Cloud. It would be
reasonably easy to write a parser and regex which special case
those exact names but disallow any other with a dash.

Signed-off-by: Adam Williamson <awilliam@redhat.com>