Drop RHEL compatibility from spec
Current version of the spec file does not work correctly on RHEL 6
without macros from EPEL. Missing python_provide causes the Provides of
the package to be incorrect.

Instead of adding more cruft to keep the compatibility, let's drop it
completely and use tito only for Fedora and EPEL builds.