Allow symbolic release versions
In order to have one compose stream track a development branch from
which release branches are created, it's useful to refer to the
in-development version symbolically (e.g., a `release_version` of `2.y`
would indicate a version 2 development stream from which 2.0, 2.1, …
configuration branches are made.

Perhaps the best description of the change is the patch for
`tests/test_common.py`.