Add additional_variants attribute to unified images
A unified image contains packages from multiple variants. It would be
really nice to be able to tell based on the metadata what those included
variants actually are.

This new attribute is only present in the output metadata if it's not
empty, and it can only be set if unified is set to True. We still allow
having unified images without this extra information.

The new attribute should be part of unique attributes types (because
there could be two different unified images (for example each would
bundle a different addon).

JIRA: COMPOSE-2916