Fix parsing composeinfo with almost conflicting UIDs
When the metadata contains variants Foo and Foo-Bar, the deserialization
logic always applies a workaround for old metadata that did not
explicitly contain parent-child relationship. Thus Foo-Bar ends up as a
child of Foo, even though they may be unrelated.

The fix is to only apply the workaround when metadata version is too
old.

JIRA: COMPOSE-3814
Signed-off-by: Lubomír Sedlář <lsedlar@redhat.com>