treeinfo: refactor deserialization code to get short name from function
We received a bug report about treeinfo.py not working on Rocky Linux
metadata, which lead towards this code. I thought it was getting to the
point of needing a bit of a refactor to use less if statements.

Comments welcome. The general approach was to just move the logic out of
the if statements and into a mapping dictionary. Hopefully this will
help make it easier for other distributions to be added as well.

I also had a thought while writing it about using a fallback mode of
trying to "guess" the short name, which is also done in e.g. anaconda
and some of the plugins based on info in the ISO root, but this would be
a significant change in functionality that likely affects other products
which rely on treeinfo data: maybe katello/foreman, satellite, others.

Thank you very much!

Best,
Neil