images: add type and formats for erofs and squashfs to back FEX
See https://fedoraproject.org/wiki/Changes/FEX ,
https://pagure.io/koji/pull-request/4219 ,
https://pagure.io/fedora-kiwi-descriptions/c/03aacd1 etc. We
seem to be growing these images in Fedora composes, so we need
a type and formats for them in productmd. These seem to be the
logical choices - all these image formats seem to be intended for
use with FEX, so I'm putting them all under the type 'fex'.

Signed-off-by: Adam Williamson <awilliam@redhat.com>