images: add more container types, move fex for alphabetical order
This adds a 'bootable-container' type. See:
https://pagure.io/fedora-iot/pungi-iot/pull-request/102#comment-210617
for some discussion here. Right now, when we build a bootc image,
its type and format are both 'ociarchive'. I don't like this;
it's redundant, and 'ociarchive' isn't really a helpful image
'type' anyway. It really just means "this is some kind of
container image", and container images can be for hugely different
purposes. I think bootc/ostree_container images having format
'ociarchive' and type 'bootable-container' is much more meaningful
and useful.

Along the same lines, why not have more types for container
images that are more meaningful about what they really are? So
this also adds 'container' that's intended to be used for "normal"
container images.

Signed-off-by: Adam Williamson <awilliam@redhat.com>